Design home section
This commit creates a home section for the website, featuring the header "My Portfolio" and the author's full name "Luis Henrique de Morais". Additionally, icons for web development languages were added using the Font Awesome library, highlighting the author's skills as a software engineer. This update establishes the home base for the website and showcases the author's professional profile.